Adding an option to hydrate collection items on collection query
If a collection query result has ‘includesItems’, then we load
collection items from the resultset instead of asking the item’s
query() method. This will allow us to specialize some of the backend
queries to optimally deliver large pages of collections, without having
to push large lists of IDs back to the backend for processing a second
time.